+++ tests/unit/configs/drivers/test_drivers_config.py
@@ -32,6 +32,11 @@ class TestDriversConfig:
             },
             "text_to_speech_driver": {"type": "DummyTextToSpeechDriver"},
             "audio_transcription_driver": {"type": "DummyAudioTranscriptionDriver"},
+            "ruleset_driver": {
+                "type": "LocalRulesetDriver",
+                "raise_not_found": True,
+                "persist_dir": None,
+            },
         }
 
     def test_from_dict(self, config):
@@ -64,6 +69,7 @@ class TestDriversConfig:
         assert Defaults.drivers_config._conversation_memory_driver is None
         assert Defaults.drivers_config._text_to_speech_driver is None
         assert Defaults.drivers_config._audio_transcription_driver is None
+        assert Defaults.drivers_config._ruleset_driver is None
 
         assert Defaults.drivers_config.prompt_driver is not None
         assert Defaults.drivers_config.image_generation_driver is not None
@@ -73,6 +79,7 @@ class TestDriversConfig:
         assert Defaults.drivers_config.conversation_memory_driver is not None
         assert Defaults.drivers_config.text_to_speech_driver is not None
         assert Defaults.drivers_config.audio_transcription_driver is not None
+        assert Defaults.drivers_config.ruleset_driver is not None
 
         assert Defaults.drivers_config._prompt_driver is not None
         assert Defaults.drivers_config._image_generation_driver is not None
@@ -82,3 +89,4 @@ class TestDriversConfig:
         assert Defaults.drivers_config._conversation_memory_driver is not None
         assert Defaults.drivers_config._text_to_speech_driver is not None
         assert Defaults.drivers_config._audio_transcription_driver is not None
+        assert Defaults.drivers_config._ruleset_driver is not None